refactor: web 패키지를 controller로 명칭 변경
